feat: complete display basket items
