+++ src/assets/styles/main.scss
@@ -26,3 +26,4 @@
 @import "./pages/productPage";
 @import "./pages/catalogPage";
 @import "./pages/aboutUs";
+@import "./pages/basketPage";
