+++ src/assets/styles/pages/_basketPage.scss
@@ -1,0 +1,19 @@
+.basket__cart-items {
+    padding: 20px 40px;
+
+    & .product {
+      display: flex;
+      border: 1px $color-active solid;
+      border-radius: 7px;
+      margin-bottom: 20px;
+
+      &__image {
+        width: 100px;
+      }
+
+      &__info {
+        margin-top: 20px;
+        margin-left: 20px;
+      }
+    }
+}
